requirements.txt
aiogram==3.4.1
telethon==1.34.0
openai==1.23.2
python-dotenv==1.0.1
tqdm==4.66.2